Apply eslint to JS tests
